pyproject.toml
[tool.poetry]
name = "api-tarpeydev"
version = "0.1.0"
description = "API for Mike Tarpey's app sandbox at tarpey.dev"
authors = ["Mike Tarpey <mike@tarpey.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
FastAPI = "^0.63.0"
odmantic = "^0.3.2"
numpy = "^1.18.4"
pandas = "^1.2.1"
requests = "^2.23.0"
scipy = "^1.4.1"
scikit-learn = "^0.24.1"
plotly = "^4.8.1"
httpx = "^0.17.0"
okta-jwt = "^1.3.5"
google-cloud-tasks = "^2.2.0"
orjson = "3.4.6"
gunicorn = "^20.0.4"
uvicorn = "^0.13.1"
aiofiles = "^0.6.0"
Jinja2 = "^2.11.2"
uvloop = "^0.14.0"
httptools = "^0.1.1"
dnspython = "^1.16.0"
SQLAlchemy = "^1.4.11"
google-cloud-storage = "^1.37.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
